requirements.txt
awesome-slugify==1.6
beautifulsoup4==4.3.2
boto==2.38.0
certifi==14.5.14
cffi==1.4.1
Collectfast==0.2.3
cryptography==1.1.2
diff-match-patch==20110725.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.8
django-admin-bootstrapped==2.4.0
django-allauth==0.23.0
django-analytical==0.21.0
django-angular==0.7.16
django-autoslug==1.8.0
django-bootstrap3==5.4.0
django-bower==5.0.1
django-braces==1.3.0
django-countries==3.3
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-easy-pdf==0.1.0
django-extensions==1.6.1
django-extra-views==0.7.1
django-filer==1.0.4
django-filter==0.7
django-floppyforms==1.6.0
django-forms-bootstrap==3.0.1
django-import-export==0.2.2
django-model-utils==2.2
django-mptt==0.7.4
django-organizations==0.5.3
django-phonenumber-field==0.7.2
django-polymorphic==0.7.2
django-secure==1.0.1
django-storages==1.1.8
django-tagging==0.3.2
django-toolbelt==0.0.1
django-wkhtmltopdf==2.0.3
django-xmlrpc==0.1.5
djangorestframework==3.3.2
drfdocs==0.0.6
easy-thumbnails==2.3
enum34==1.1.1
gunicorn==19.1.1
html5lib==0.9999999
idna==2.0
ipaddress==1.0.15
ipdb==0.8
ipython==2.2.0
Jinja2==2.8
Markdown==2.4.1
MarkupSafe==0.23
oauthlib==1.0.3
opbeat==3.2.1
phonenumbers==7.0.7
Pillow==3.1.1
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
pyparsing==2.0.2
PyPDF2==1.25.1
pystache==0.5.3
python-dateutil==2.4.2
python-nvd3==0.14.2
python-openid==2.2.5
python-slugify==1.1.4
pytz==2014.4
pyzmq==14.6.0
regex==2014.8.28
reportlab==3.3.0
requests==2.7.0
requests-oauthlib==0.5.0
six==1.10.0
smtpapi==0.1.2
sqlparse==0.1.11
static==1.0.2
static3==0.5.1
tablib==0.9.11
Unidecode==0.4.18
unittest2==0.5.1
waitress==0.8.9
Werkzeug==0.9.6
wkhtmltopdf==0.2
xhtml2pdf==0.0.6
